[all] Added stats functions. Fixed pipelining issues by appending the readyQueue to also store time when a readying token came in. We can then later check in acknowledgeToken if enough time has passed since the token came in that we don't have to add frontendLatency. Builds and Runs

+++ cgra_defs.h
@@ -19,7 +19,7 @@
 using std::cout; // TODO (by): Remove when done with debugging?
 
 // TODO (nzb): Remove when merged.
-DERIVE_STRONGER_INT(Cycles, uint64_t, cycles)
+DERIVE_STRONGER_INT(Cycles, int32_t, cycles)
 const size_t MAX_ARGS = 6;
 const size_t MAX_ARG_BYTES = MAX_ARGS * sizeof(uint64_t);
 DERIVE_STRONGER_INT(ProcIdx, int32_t, pid);
